refactor: GardensContainer 리팩토링
